+++ djangocms_alias/admin.py
@@ -57,7 +57,7 @@ class AliasAdmin(admin.ModelAdmin):
     list_display = ['name', 'category']
     list_filter = ['site', 'category']
     fields = ('category', 'site')
-    readonly_fields = ('static_code',)
+    readonly_fields = ('static_code', )
 
     def get_urls(self):
         return urlpatterns + super().get_urls()